init :: 파일구조 분리중인 path_server 임시 저장
