+++ src/config/db/types/schedulerSettings.ts
