+++ src/infrastructure/ai/openaiClient.ts
