+++ src/jobs/schedulerSettings.ts
@@ -2,7 +2,7 @@
 
 import { RowDataPacket } from "mysql2";
 import { getDatabase } from "../config/db";
-import { SchedulerSettings } from "../modules/types/schedulerSettings";
+import { SchedulerSettings } from "../config/db/types/schedulerSettings";
 
 /**
  * DB에서 마지막 스케줄러 실행 날짜(last_run_date)를 가져옵니다.
