+++ src/modules/auth/swaggerAuthPaths.ts
