+++ src/modules/concepts/repositories/conceptsRepository.ts
@@ -1,0 +1,5 @@
+export class ConceptsRepository {
+  static async findByid(concepts: string): Promise<void> {}
+
+  static async findByPage(page: number): Promise<void> {}
+}
