--- src/modules/types/financialStatementChatRoom.ts
@@ -1,8 +1,0 @@
-export interface FinancialStatementChatRoom {
-  id: string;
-  user_id: string;
-  name: string;
-  description: string | null;
-  created_at: Date;
-  updated_at: Date;
-}
