+++ src/modules/user/types/dto/response/user-dashboard.response.ts
@@ -1,0 +1,75 @@
+// ============================================
+// Response Types
+// ============================================
+
+/**
+ * 3.1.1 목표 진척도 (온보딩 질문)
+ * 3.1.2 목표 진척도 (재무설계)
+ */
+export interface GoalProgressItem {
+  id: string; // 목표 ID
+  name: string; // 목표명
+  description: string | null; // 목표 설명
+  targetAmount: number; // 목표 금액
+  currentAmount: number; // 현재 금액
+  progressPercentage: number; // 진척도 (%)
+  monthlyContribution: number; // 월 기여도
+  startDate: string | null; // 시작 날짜
+  endDate: string | null; // 종료 날짜
+}
+
+/**
+ * 3.1.3 월별 수입·지출 데이터
+ */
+export interface MonthlyFinanceItem {
+  year: number;
+  month: number;
+  income: number;
+  expense: number;
+}
+
+/**
+ * 3.1.4 재무설계표 채팅방
+ */
+export interface FinancialStatementChatRoom {
+  id: string;
+  name: string;
+  description: string | null;
+  createdAt: string;
+  updatedAt: string | null;
+}
+
+/**
+ * 3.1.4 일반 채팅방 (RAG 기반)
+ */
+export interface RAGChatRoom {
+  id: string;
+  name: string;
+  description: string | null;
+  memoryCount: number;
+  createdAt: string;
+  updatedAt: string | null;
+}
+
+/**
+ * 유저 대시보드 화면 응답
+ */
+export interface UserDashboardResponse {
+  // 3.1.1 온보딩 질문에서 받은 목표 진척도
+  onboardingGoals: GoalProgressItem[];
+
+  // 3.1.2 재무설계에서 생긴 목표 진척도
+  financialPlansGoals: GoalProgressItem[];
+
+  // 3.1.3 달별 수입·지출 그래프 (연도별, 생성 순으로 정렬)
+  monthlyFinances: MonthlyFinanceItem[];
+
+  // 3.1.4 채팅 목록
+  chatRooms: {
+    // 재무설계표 채팅 (최근 생성 순)
+    financialStatementChats: FinancialStatementChatRoom[];
+
+    // RAG 기반 채팅 (최근 생성 순)
+    ragChats: RAGChatRoom[];
+  };
+}
