+++ src/modules/user/types/dto/response/user-info.response.ts
@@ -1,0 +1,11 @@
+//유저 정보 응답
+export interface UserInfoResponse {
+  userId: string;
+  email: string;
+  name: string;
+  hasLoan: boolean;
+  hasStock: boolean;
+  recentPlanDate: Date | null;
+  createdAt: Date;
+  updatedAt: Date;
+}
